Size is not grandeur

+++ 0x00-hello_world/6-size.c
@@ -12,7 +12,7 @@ int main(void)
 	char c;
 	int i;
 	long li;
-	long lli;
+	long long lli;
 	float f;
 
 	printf("Size of a char: %ld byte(s)\n", sizeof(c));
